Bugfux

+++ ptypy/experiment/I08.py
@@ -192,7 +192,7 @@ class I08_Scan(ptypy.core.data.PtyScan):
             ix[i] = int(np.mod(indices[i], self.common.scan_dimensions[1]))  # find the remainder - works out the column
             iy[i] = int(indices[i]//self.common.scan_dimensions[0])  # works out the row
             raw[i] = (io.h5read(self.nxs_filename, key, slice=(ix[i], iy[i]))[key].astype(np.float32)-self.common.dark)/ (self.common.flat)  #-self.common.dark) # load in the data and convert type
-            weights[i] = float(raw[i] >= 0.)
+            weights[i] = (raw[i] >= 0.).astype(float)
         return raw, pos, weights
 
 #io.h5read('somefile.h5', data[(slice1, slice2)]
